refactor internal/backend to split out different areas of functionality
