+++ api/doc.go
@@ -1,4 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+// Package api provides types and interfaces that can be used by
+// client-side software to communicate with finch daemon.
+// It currently implements the Docker API Spec v1.43.
 package api
